+++ Tests/KeyboardKitTests/Actions/KeyboardActionTests.swift
@@ -15,6 +15,47 @@ class KeyboardActionTests: QuickSpec {
     
     override func spec() {
         
+        var actions: [KeyboardAction]!
+        
+        var expected: [KeyboardAction]! {
+            didSet {
+                unexpected = actions
+                expected.forEach { action in
+                    unexpected.removeAll { $0 == action }
+                }
+            }
+        }
+        
+        var unexpected: [KeyboardAction]!
+        
+        beforeEach {
+            actions = [
+                .none,
+                .dismissKeyboard,
+                .capsLock,
+                .character(""),
+                .command,
+                .custom(name: ""),
+                .escape,
+                .function,
+                .image(description: "", keyboardImageName: "", imageName: ""),
+                .keyboardType(.email),
+                .moveCursorBackward,
+                .moveCursorForward,
+                .newLine,
+                .nextKeyboard,
+                .option,
+                .shift,
+                .shiftDown,
+                .space,
+                .systemImage(description: "", keyboardImageName: "", imageName: ""),
+                .tab,
+            ]
+            
+            expected = []
+            unexpected = []
+        }
+        
         describe("is delete action") {
             
             func result(for action: KeyboardAction) -> Bool {
@@ -22,28 +63,9 @@ class KeyboardActionTests: QuickSpec {
             }
             
             it("is true for some actions") {
-                expect(result(for: .backspace)).to(beTrue())
-                
-                expect(result(for: .none)).to(beFalse())
-                expect(result(for: .dismissKeyboard)).to(beFalse())
-                expect(result(for: .capsLock)).to(beFalse())
-                expect(result(for: .character(""))).to(beFalse())
-                expect(result(for: .command)).to(beFalse())
-                expect(result(for: .custom(name: ""))).to(beFalse())
-                expect(result(for: .escape)).to(beFalse())
-                expect(result(for: .function)).to(beFalse())
-                expect(result(for: .image(description: "", keyboardImageName: "", imageName: ""))).to(beFalse())
-                expect(result(for: .keyboardType(.email))).to(beFalse())
-                expect(result(for: .moveCursorBackward)).to(beFalse())
-                expect(result(for: .moveCursorForward)).to(beFalse())
-                expect(result(for: .newLine)).to(beFalse())
-                expect(result(for: .nextKeyboard)).to(beFalse())
-                expect(result(for: .option)).to(beFalse())
-                expect(result(for: .shift)).to(beFalse())
-                expect(result(for: .shiftDown)).to(beFalse())
-                expect(result(for: .space)).to(beFalse())
-                expect(result(for: .systemImage(description: "", keyboardImageName: "", imageName: ""))).to(beFalse())
-                expect(result(for: .tab)).to(beFalse())
+                expected = [.backspace]
+                expected.forEach { expect(result(for: $0)).to(beTrue()) }
+                unexpected.forEach { expect(result(for: $0)).to(beFalse()) }
             }
         }
         
@@ -54,30 +76,15 @@ class KeyboardActionTests: QuickSpec {
             }
             
             it("is true for some actions") {
-                expect(result(for: .character(""))).to(beTrue())
-                expect(result(for: .emoji(""))).to(beTrue())
-                expect(result(for: .image(description: "", keyboardImageName: "", imageName: ""))).to(beTrue())
-                expect(result(for: .space)).to(beTrue())
-                expect(result(for: .systemImage(description: "", keyboardImageName: "", imageName: ""))).to(beTrue())
-                
-                expect(result(for: .none)).to(beFalse())
-                expect(result(for: .backspace)).to(beFalse())
-                expect(result(for: .dismissKeyboard)).to(beFalse())
-                expect(result(for: .capsLock)).to(beFalse())
-                expect(result(for: .command)).to(beFalse())
-                expect(result(for: .custom(name: ""))).to(beFalse())
-                expect(result(for: .emojiCategory(.activities))).to(beFalse())
-                expect(result(for: .escape)).to(beFalse())
-                expect(result(for: .function)).to(beFalse())
-                expect(result(for: .keyboardType(.email))).to(beFalse())
-                expect(result(for: .moveCursorBackward)).to(beFalse())
-                expect(result(for: .moveCursorForward)).to(beFalse())
-                expect(result(for: .newLine)).to(beFalse())
-                expect(result(for: .nextKeyboard)).to(beFalse())
-                expect(result(for: .option)).to(beFalse())
-                expect(result(for: .shift)).to(beFalse())
-                expect(result(for: .shiftDown)).to(beFalse())
-                expect(result(for: .tab)).to(beFalse())
+                expected = [
+                    .character(""),
+                    .emoji(""),
+                    .image(description: "", keyboardImageName: "", imageName: ""),
+                    .space,
+                    .systemImage(description: "", keyboardImageName: "", imageName: "")
+                ]
+                expected.forEach { expect(result(for: $0)).to(beTrue()) }
+                unexpected.forEach { expect(result(for: $0)).to(beFalse()) }
             }
         }
         
@@ -88,30 +95,25 @@ class KeyboardActionTests: QuickSpec {
             }
             
             it("is true for some actions") {
-                expect(result(for: .backspace)).to(beTrue())
-                expect(result(for: .dismissKeyboard)).to(beTrue())
-                expect(result(for: .capsLock)).to(beTrue())
-                expect(result(for: .command)).to(beTrue())
-                expect(result(for: .escape)).to(beTrue())
-                expect(result(for: .function)).to(beTrue())
-                expect(result(for: .keyboardType(.email))).to(beTrue())
-                expect(result(for: .moveCursorBackward)).to(beTrue())
-                expect(result(for: .moveCursorForward)).to(beTrue())
-                expect(result(for: .newLine)).to(beTrue())
-                expect(result(for: .nextKeyboard)).to(beTrue())
-                expect(result(for: .option)).to(beTrue())
-                expect(result(for: .shift)).to(beTrue())
-                expect(result(for: .shiftDown)).to(beTrue())
-                expect(result(for: .tab)).to(beTrue())
-                
-                expect(result(for: .none)).to(beFalse())
-                expect(result(for: .character(""))).to(beFalse())
-                expect(result(for: .custom(name: ""))).to(beFalse())
-                expect(result(for: .emoji(""))).to(beFalse())
-                expect(result(for: .emojiCategory(.foods))).to(beFalse())
-                expect(result(for: .image(description: "", keyboardImageName: "", imageName: ""))).to(beFalse())
-                expect(result(for: .space)).to(beFalse())
-                expect(result(for: .systemImage(description: "", keyboardImageName: "", imageName: ""))).to(beFalse())
+                expected = [
+                    .backspace,
+                    .dismissKeyboard,
+                    .capsLock,
+                    .command,
+                    .escape,
+                    .function,
+                    .keyboardType(.email),
+                    .moveCursorBackward,
+                    .moveCursorForward,
+                    .newLine,
+                    .nextKeyboard,
+                    .option,
+                    .shift,
+                    .shiftDown,
+                    .tab
+                ]
+                expected.forEach { expect(result(for: $0)).to(beTrue()) }
+                unexpected.forEach { expect(result(for: $0)).to(beFalse()) }
             }
         }
         
@@ -122,30 +124,9 @@ class KeyboardActionTests: QuickSpec {
             }
             
             it("is defined for some actions") {
-                expect(result(for: .shift)).toNot(beNil())
-                expect(result(for: .space)).toNot(beNil())
-                
-                expect(result(for: .none)).to(beNil())
-                expect(result(for: .backspace)).to(beNil())
-                expect(result(for: .capsLock)).to(beNil())
-                expect(result(for: .character("a"))).to(beNil())
-                expect(result(for: .dismissKeyboard)).to(beNil())
-                expect(result(for: .command)).to(beNil())
-                expect(result(for: .custom(name: ""))).to(beNil())
-                expect(result(for: .emoji(""))).to(beNil())
-                expect(result(for: .emojiCategory(.foods))).to(beNil())
-                expect(result(for: .escape)).to(beNil())
-                expect(result(for: .function)).to(beNil())
-                expect(result(for: .image(description: "", keyboardImageName: "", imageName: ""))).to(beNil())
-                expect(result(for: .keyboardType(.email))).to(beNil())
-                expect(result(for: .moveCursorBackward)).to(beNil())
-                expect(result(for: .moveCursorForward)).to(beNil())
-                expect(result(for: .newLine)).to(beNil())
-                expect(result(for: .nextKeyboard)).to(beNil())
-                expect(result(for: .option)).to(beNil())
-                expect(result(for: .shiftDown)).to(beNil())
-                expect(result(for: .systemImage(description: "", keyboardImageName: "", imageName: ""))).to(beNil())
-                expect(result(for: .tab)).to(beNil())
+                expected = [.shift, .space]
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
             }
         }
         
@@ -157,27 +138,7 @@ class KeyboardActionTests: QuickSpec {
             }
             
             it("is defined for all actions that has a standard tap action") {
-                expect(result(for: .none)).to(beTrue())
-                expect(result(for: .backspace)).to(beTrue())
-                expect(result(for: .capsLock)).to(beTrue())
-                expect(result(for: .character("a"))).to(beTrue())
-                expect(result(for: .dismissKeyboard)).to(beTrue())
-                expect(result(for: .command)).to(beTrue())
-                expect(result(for: .custom(name: ""))).to(beTrue())
-                expect(result(for: .escape)).to(beTrue())
-                expect(result(for: .function)).to(beTrue())
-                expect(result(for: .image(description: "", keyboardImageName: "", imageName: ""))).to(beTrue())
-                expect(result(for: .keyboardType(.email))).to(beTrue())
-                expect(result(for: .moveCursorBackward)).to(beTrue())
-                expect(result(for: .moveCursorForward)).to(beTrue())
-                expect(result(for: .newLine)).to(beTrue())
-                expect(result(for: .nextKeyboard)).to(beTrue())
-                expect(result(for: .option)).to(beTrue())
-                expect(result(for: .shift)).to(beTrue())
-                expect(result(for: .shiftDown)).to(beTrue())
-                expect(result(for: .space)).to(beTrue())
-                expect(result(for: .systemImage(description: "", keyboardImageName: "", imageName: ""))).to(beTrue())
-                expect(result(for: .tab)).to(beTrue())
+                actions.forEach { expect(result(for: $0)).to(beTrue()) }
             }
         }
         
@@ -188,30 +149,22 @@ class KeyboardActionTests: QuickSpec {
             }
             
             it("is defined for some actions") {
-                expect(result(for: .backspace)).toNot(beNil())
-                expect(result(for: .character("a"))).toNot(beNil())
-                expect(result(for: .dismissKeyboard)).toNot(beNil())
-                expect(result(for: .emoji(""))).toNot(beNil())
-                expect(result(for: .keyboardType(.email))).toNot(beNil())
-                expect(result(for: .moveCursorBackward)).toNot(beNil())
-                expect(result(for: .moveCursorForward)).toNot(beNil())
-                expect(result(for: .newLine)).toNot(beNil())
-                expect(result(for: .shift)).toNot(beNil())
-                expect(result(for: .shiftDown)).toNot(beNil())
-                expect(result(for: .space)).toNot(beNil())
-                expect(result(for: .tab)).toNot(beNil())
-                
-                expect(result(for: .none)).to(beNil())
-                expect(result(for: .capsLock)).to(beNil())
-                expect(result(for: .command)).to(beNil())
-                expect(result(for: .custom(name: ""))).to(beNil())
-                expect(result(for: .emojiCategory(.foods))).to(beNil())
-                expect(result(for: .escape)).to(beNil())
-                expect(result(for: .function)).to(beNil())
-                expect(result(for: .image(description: "", keyboardImageName: "", imageName: ""))).to(beNil())
-                expect(result(for: .nextKeyboard)).to(beNil())
-                expect(result(for: .option)).to(beNil())
-                expect(result(for: .systemImage(description: "", keyboardImageName: "", imageName: ""))).to(beNil())
+                expected = [
+                    .backspace,
+                    .character(""),
+                    .dismissKeyboard,
+                    .emoji(""),
+                    .keyboardType(.email),
+                    .moveCursorBackward,
+                    .moveCursorForward,
+                    .newLine,
+                    .shift,
+                    .shiftDown,
+                    .space,
+                    .tab
+                ]
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
             }
         }
         
@@ -222,30 +175,31 @@ class KeyboardActionTests: QuickSpec {
             }
             
             it("is defined for some actions") {
-                expect(result(for: .backspace)).toNot(beNil())
-                
-                expect(result(for: .none)).to(beNil())
-                expect(result(for: .capsLock)).to(beNil())
-                expect(result(for: .character("a"))).to(beNil())
-                expect(result(for: .dismissKeyboard)).to(beNil())
-                expect(result(for: .command)).to(beNil())
-                expect(result(for: .custom(name: ""))).to(beNil())
-                expect(result(for: .emoji(""))).to(beNil())
-                expect(result(for: .emojiCategory(.foods))).to(beNil())
-                expect(result(for: .escape)).to(beNil())
-                expect(result(for: .function)).to(beNil())
-                expect(result(for: .image(description: "", keyboardImageName: "", imageName: ""))).to(beNil())
-                expect(result(for: .keyboardType(.email))).to(beNil())
-                expect(result(for: .moveCursorBackward)).to(beNil())
-                expect(result(for: .moveCursorForward)).to(beNil())
-                expect(result(for: .newLine)).to(beNil())
-                expect(result(for: .nextKeyboard)).to(beNil())
-                expect(result(for: .option)).to(beNil())
-                expect(result(for: .shift)).to(beNil())
-                expect(result(for: .shiftDown)).to(beNil())
-                expect(result(for: .space)).to(beNil())
-                expect(result(for: .systemImage(description: "", keyboardImageName: "", imageName: ""))).to(beNil())
-                expect(result(for: .tab)).to(beNil())
+                expected = [.backspace]
+                expected.forEach { expect(result(for: $0)).toNot(beNil()) }
+                unexpected.forEach { expect(result(for: $0)).to(beNil()) }
+            }
+        }
+        
+        describe("system font") {
+            
+            it("is defined for all actions") {
+                actions.forEach {
+                    expect($0.systemFont).to(equal(UIFont.preferredFont(forTextStyle: $0.systemTextStyle)))
+                }
+            }
+        }
+        
+        describe("system text style") {
+            
+            it("is custom for some actions, but defined for all") {
+                actions.forEach {
+                    if $0 == .emoji("") {
+                        expect($0.systemTextStyle).to(equal(.title1))
+                    } else {
+                        expect($0.systemTextStyle).to(equal(.title2))
+                    }
+                }
             }
         }
     }